% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_exp-data.R
\docType{data}
\name{gene_exp}
\alias{gene_exp}
\title{Gene Expression Dataset for miknn}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 11093 rows and 4 columns.}
\usage{
data(gene_exp)
}
\description{
Gene Expression Dataset for miknn
}
\examples{
data("gene_exp", package = "miknn")
}
\keyword{datasets}
