% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi_knn_fast.R
\name{mi_knn.fast}
\alias{mi_knn.fast}
\title{Faster Mutual Information Calculation (experimental)}
\usage{
mi_knn.fast(df, var.d, var.c, k, output = "default")
}
\arguments{
\item{df}{a data.frame object}

\item{var.d}{the name of the discrete variable in quotations}

\item{var.c}{the name of the continuouse variable in quotations}

\item{k}{the number of neighbors for MI calculation.}

\item{output}{set to 'raw' to get the I calculated for each point, or 'global' for globa MI.}
}
\description{
Faster Mutual Information Calculation (experimental)
}
