% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi_knn.R
\name{mi_knn}
\alias{mi_knn}
\title{Mutual Information Calculation}
\usage{
mi_knn(dt, var.d, var.c, k = NULL, warnings = TRUE, FORCE = TRUE,
  global = TRUE, quite = FALSE)
}
\arguments{
\item{dt}{a data.frame object}

\item{var.d}{the name of the discrete variable in quotations}

\item{var.c}{the name of the continuouse variable in quotations}

\item{k}{the number of neighbors for MI calculation.}

\item{warnings}{set to FALSE to hide warnings}

\item{FORCE}{set to TRUE continues despite not using an optimum k}

\item{global}{set to FALSE to get specific MI}

\item{quite}{set to TRUE to prevent messages being displayed to the console}
}
\description{
Mutual Information Calculation
}
\examples{
library(miknn)

#TCGA data
data("gene_exp",package = "miknn")
gene_exp

# basic example using k = 10
mi_knn(dt = gene_exp, var.d = "project", k = 10, var.c = "CCND1")

# the function can decide the optimum k
mi_knn(dt = gene_exp, var.d = "project", var.c = "CCND1")

# set global = FALSE to get the specific MI for each levels,  
mi_knn(dt = gene_exp, var.d = "project", var.c = "CCND1", global = FALSE)

# generating a plot for specific MI
res <- mi_knn(dt = gene_exp, var.d = "project", var.c = "CCND1", global = FALSE)
plot(res)

}
