% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi_kopt.R
\name{mi_kopt}
\alias{mi_kopt}
\title{Determine the Optimum K}
\usage{
mi_kopt(dt, var.d, var.c, k.series, global = TRUE)
}
\arguments{
\item{dt}{a data.frame object}

\item{var.d}{the name of the discrete variable in quotations}

\item{var.c}{the name of the continuouse variable in quotations}

\item{k.series}{a vector of integers to test}

\item{global}{set as TRUE by default. Set to FALSE for specific MI}
}
\description{
Determine the Optimum K
}
\examples{
library(miknn)

# determining I for a series of k
k.series <- 3:40
mi_kopt(iris, "Species","Sepal.Length", k.series, global = FALSE)

}
