% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi_kopt.R
\name{plot.mi_kopt}
\alias{plot.mi_kopt}
\title{Plot mi_kopt}
\usage{
\method{plot}{mi_kopt}(x, ...)
}
\arguments{
\item{x}{result from mi_kopt}

\item{...}{For the "function" method of plot, ... can include any of the other arguments of curve, except expr.}
}
\description{
Plot mi_kopt
}
