% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi_knn.R
\name{plot.specific_mi}
\alias{plot.specific_mi}
\title{plot.specific_mi}
\usage{
\method{plot}{specific_mi}(x, ...)
}
\arguments{
\item{x}{a result from mi_knn(dt, var.d,var.d, global = FALSE)}

\item{...}{For the "function" method of plot, ... can include any of the other arguments of curve, except expr.}
}
\description{
plot.specific_mi
}
