% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi_knn.R
\name{print.global_mi}
\alias{print.global_mi}
\title{print.global_mi}
\usage{
\method{print}{global_mi}(x, ...)
}
\arguments{
\item{x}{an mi result}

\item{...}{For the "function" method of plot, ... can include any of the other arguments of curve, except expr.}
}
\description{
print.global_mi
}
