% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi_kopt.R
\name{mi_kopt}
\alias{mi_kopt}
\title{Determine the Optimum K}
\usage{
mi_kopt(dt, var.d, var.c, global = TRUE, returnK = FALSE,
  progress = TRUE, k.series = NULL)
}
\arguments{
\item{dt}{a data.frame object}

\item{var.d}{the name of the discrete variable in quotations}

\item{var.c}{the name of the continuouse variable in quotations}

\item{global}{set as TRUE by default. Set to FALSE for specific MI}

\item{returnK}{set as TRUE to return the 'optimal' k}

\item{progress}{whether to show a progress bar (DEFAULT=TRUE)}

\item{k.series}{the range of integers to scan through (DEFAULT=1:100)}
}
\description{
Determine the Optimum K
}
\examples{
library(miknn)

# determining I for a series of k
mi_kopt(iris, "Species","Sepal.Length", global = TRUE)

}
