% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi_kopt.R
\name{mi_kopt}
\alias{mi_kopt}
\title{Determine the Optimum K}
\usage{
mi_kopt(dt, var.d, var.c, global = TRUE, returnK = FALSE)
}
\arguments{
\item{dt}{a data.frame object}

\item{var.d}{the name of the discrete variable in quotations}

\item{var.c}{the name of the continuouse variable in quotations}

\item{global}{set as TRUE by default. Set to FALSE for specific MI}

\item{k.series}{a vector of integers to test}
}
\description{
Determine the Optimum K
}
\examples{
library(miknn)

# determining I for a series of k
mi_kopt(iris, "Species","Sepal.Length", global = FALSE)

}
